Extract JSOptionalTest212.traitWithOptionalFunction() in a separate file.
Scala 3 cannot infer the parameter type yet when using true union
types instead of Scala.js' pseudo union types. We therefore extract
that test in a separate file so that Scala 3 can blacklist it while
preserving the other tests in the file.
